refactor(compiler): refactor lexer
